+++ installers/mac/build.sh
@@ -92,7 +92,11 @@ DEPENDENCIES_DIR="$COMPONENTS_DIR/dependencies"
 rm -rf "$DEPENDENCIES_DIR"
 mkdir -p "$DEPENDENCIES_DIR"
 unzip -o "$JRE_ZIP" -d "$DEPENDENCIES_DIR"
-JRE_FOLDER=$(unzip -Z1 "$JRE_ZIP" | head -1 | cut -d/ -f1)
+JRE_FOLDER=$(unzip -Z1 "$JRE_ZIP" | awk -F/ '{print $1}' | sort -u | grep -v '^$' | head -1)
+if [ -z "$JRE_FOLDER" ]; then
+    print_error "Could not determine JRE folder from zip"
+    exit 1
+fi
 
 rm -rf "$BALLERINA_UNZIPPED_PATH"
 rm -rf "$BALLERINA_TEMP"
@@ -120,8 +124,8 @@ chmod +x "$ICP_TARGET/bin"/*
 ICP_SCRIPT="$ICP_TARGET/bin/icp.sh"
 if [ -f "$ICP_SCRIPT" ]; then
     print_info "Modifying icp.sh to use JRE from dependencies ($JRE_FOLDER)"
-    # Replace all java instances with the full path to the JRE java
-    sed -i '' "s|java|\"\$SCRIPT_DIR\"/../../dependencies/$JRE_FOLDER/bin/java|g" "$ICP_SCRIPT"
+    # Replace standalone 'java' invocations with the full path to the JRE java (word-boundary match)
+    sed -i '' -E "s|[[:<:]]java[[:>:]]|\"\$SCRIPT_DIR\"/../../dependencies/$JRE_FOLDER/bin/java|g" "$ICP_SCRIPT"
 fi
